requirements..txt
faiss-cpu==1.7.4
google-generativeai
Pillow
requests
torch
transformers